Add BathroomAlert Feature
Add the ability for users to create an sms alert to be notified when a
bathroom is available.

Add the model, migrations, REST API and tests.

Add app config to import signals.py at the proper time.

Add signals.py to handle the notification of alerts. When an
EnvironmentEvent is saved and the bathroom is available notify.